file created though not populating

+++ src/core/settings.py
@@ -8,6 +8,8 @@ import gzip
 import shutil
 import datetime
 import schedule
+import pathlib
+from logging.handlers import TimedRotatingFileHandler
 
 from pathlib import Path
 
@@ -354,29 +356,50 @@ LOGGING = {
 
 current_date = datetime.date.today()
 
-absolute_path_to_logs = os.path.abspath("src/logs")
-file_path = os.path.abspath("src/logs/debug.log")
-new_file_path = os.path.join("src/past_logs", str(current_date) + ".gz")
-# print(absolute_path_to_logs)
-# print(file_path)
+# logging.basicConfig(filename='scheduler.log')
+# schedule_logger = logging.getLogger('schedule')
+# schedule_logger.setLevel(level=logging.DEBUG)
 
-def compress_old_logs():
-    """
-    Moves and compresses rotated log files into 'past_logs' directory."
-    """
-    # Compress log file
-    with open(file_path, "rb") as f_in, gzip.open(new_file_path, "wb") as f_out:
-        print("this works")
-        shutil.copyfileobj(f_in, f_out)
 
-    # Clear all the files in src/logs log file after compression
-    for file in os.listdir(absolute_path_to_logs):
-        logs_file_path = os.path.join(absolute_path_to_logs, file)
-        if not os.path.exists(logs_file_path):
-            open(file, 'w').close()
-            print(f"File added: {logs_file_path}")
-        open(logs_file_path, 'r+').truncate(0)
+class GzipTimedRotatingFileHandler(TimedRotatingFileHandler):
+    def doRollover(self):
+        super().doRollover()
+        # log_file = self.baseFilename
+        if os.path.exists(file_path):
+            with open(file_path, 'rb') as f_in, gzip.open(log_file_name, "wb") as f_out:
+                print(f_in)
+                shutil.copyfileobj(f_in, f_out)
 
-    print(f"Compressed and moved: {new_file_path}")
+            # Clear all the files in src/logs log file after compression
+            for file in os.listdir(absolute_path_to_logs):
+                logs_file_path = os.path.join(absolute_path_to_logs, file)
+                if not os.path.exists(logs_file_path):
+                    open(file, 'w').close()
+                    print(f"File added: {logs_file_path}")
+                open(logs_file_path, 'r+').truncate(0)
 
-schedule.every(5).minutes.do(compress_old_logs)
+
+# Configure logging
+# logger = logging.getLogger(__name__)
+# logger.setLevel(logging.DEBUG)
+
+
+logging.basicConfig(filename='scheduler.log')
+schedule_logger = logging.getLogger('schedule')
+schedule_logger.setLevel(level=logging.DEBUG)
+
+MODULE_PATH = pathlib.Path(__file__).parent.parent.resolve()
+absolute_path_to_logs = os.path.join(MODULE_PATH, "logs")
+file_path = os.path.join(absolute_path_to_logs, "debug.log")
+log_file_name = os.path.join(MODULE_PATH, "past_logs", str(current_date) + ".gz")
+
+# handler = GzipTimedRotatingFileHandler(log_file_name, when='S', interval=5, backupCount=5)
+handler = GzipTimedRotatingFileHandler(log_file_name, when='M', interval=1)
+# handler = GzipTimedRotatingFileHandler(log_file_name, when='midnight')
+
+# Set the log message format
+# formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
+# handler.setFormatter(formatter)
+
+# # Add the handler to the logger
+# logger.addHandler(handler)
